Move context to a utils module
